// packages/shared-logic/src/types.ts
export interface Review {
  id: number;
  userId: number;
  username: string;
  avatar: string;
  rating: number;
  comment: string;
  date: string;
  adminReply?: string;
  helpfulCount?: number;
}

export interface Service {
  id: number;
  subCategoryId: number; // Links to a sub-category
  ownerId?: number; // Link to the AppUser who owns this service
  name: string;
  images: string[];
  address: string;
  phone: string[];
  whatsapp: string[];
  about: string;
  rating: number; // This would typically be calculated
  reviews: Review[];
  facebookUrl?: string;
  instagramUrl?: string;
  locationUrl?: string; // Optional Google Maps link
  workingHours?: string; // Optional text for working hours
  isFavorite: boolean;
  views: number;
  creationDate: string;
}

export interface SubCategory {
  id: number;
  name: string;
}

export interface Category {
  id: number;
  name: string;
  icon: string; // The name of the icon component
  subCategories: SubCategory[];
}

export interface News {
  id: number;
  title: string;
  content: string;
  imageUrl: string;
  date: string;
  author: string;
  externalUrl?: string;
  views: number;
}

export interface Notification {
  id: number;
  title: string;
  content: string;
  imageUrl?: string;
  externalUrl?: string;
  serviceId?: number; // Link to a service
  startDate: string;
  endDate: string;
}

export interface Advertisement {
  id: number;
  title: string;
  imageUrl: string;
  serviceId?: number;
  externalUrl?: string;
  startDate: string;
  endDate: string;
}

export interface Property {
  id: number;
  title: string;
  description: string;
  images: string[];
  location: {
    address: string;
  };
  type: 'sale' | 'rent';
  price: number;
  contact: {
    name: string;
    phone: string;
  };
  amenities: string[];
  views: number;
  creationDate: string;
}